Query helpers added
